Expand max number of overworlds to 65535

+++ src/cable_car.c
@@ -792,7 +792,7 @@ static void CreateCableCarSprites(void)
     u8 i;
 
     u16 rval = Random();
-    u8 hikerGraphicsIds[4] = {
+    u16 hikerGraphicsIds[4] = {
         OBJ_EVENT_GFX_HIKER,
         OBJ_EVENT_GFX_CAMPER,
         OBJ_EVENT_GFX_PICNICKER,
