+++ src/field_player_avatar.c
@@ -218,7 +218,7 @@ static bool8 (*const sArrowWarpMetatileBehaviorChecks[])(u8) =
     [DIR_EAST - 1]  = MetatileBehavior_IsEastArrowWarp,
 };
 
-static const u8 sPlayerAvatarGfxIds[][2] =
+static const u16 sPlayerAvatarGfxIds[][2] =
 {
     [PLAYER_AVATAR_STATE_NORMAL]     = {OBJ_EVENT_GFX_BRENDAN_NORMAL,     OBJ_EVENT_GFX_MAY_NORMAL},
     [PLAYER_AVATAR_STATE_BIKE]       = {OBJ_EVENT_GFX_BRENDAN_BIKE,       OBJ_EVENT_GFX_MAY_BIKE},
@@ -226,7 +226,7 @@ static const u8 sPlayerAvatarGfxIds[][2] =
     [PLAYER_AVATAR_STATE_UNDERWATER] = {OBJ_EVENT_GFX_BRENDAN_UNDERWATER, OBJ_EVENT_GFX_MAY_UNDERWATER},
 };
 
-static const u8 sPlayerAvatarAnimGfxIds[][2] =
+static const u16 sPlayerAvatarAnimGfxIds[][2] =
 {
     [PLAYER_AVATAR_GFX_FIELD_MOVE] = {OBJ_EVENT_GFX_BRENDAN_FIELD_MOVE, OBJ_EVENT_GFX_MAY_FIELD_MOVE},
     [PLAYER_AVATAR_GFX_FISHING]    = {OBJ_EVENT_GFX_BRENDAN_FISHING,    OBJ_EVENT_GFX_MAY_FISHING},
@@ -1213,14 +1213,14 @@ void StopPlayerAvatar(void)
     }
 }
 
-u8 GetPlayerAvatarGraphicsIdByStateIdAndGender(u8 state, u8 gender)
+u16 GetPlayerAvatarGraphicsIdByStateIdAndGender(u8 state, u8 gender)
 {
     return sPlayerAvatarGfxIds[state][gender];
 }
 
-u8 GetLinkPlayerAvatarGraphicsIdByGender(u8 version, u8 gender)
+u16 GetLinkPlayerAvatarGraphicsIdByGender(u8 version, u8 gender)
 {
-    u8 gfxId = sPlayerAvatarGfxIds[PLAYER_AVATAR_STATE_NORMAL][gender];
+    u16 gfxId = sPlayerAvatarGfxIds[PLAYER_AVATAR_STATE_NORMAL][gender];
     switch (GetVersionId(version))
     {
     case 2:
@@ -1233,7 +1233,7 @@ u8 GetLinkPlayerAvatarGraphicsIdByGender(u8 version, u8 gender)
     return gfxId;
 }
 
-u8 GetPlayerAvatarGraphicsIdByStateId(u8 state)
+u16 GetPlayerAvatarGraphicsIdByStateId(u8 state)
 {
     return GetPlayerAvatarGraphicsIdByStateIdAndGender(state, gSaveBlock2Ptr->playerGender);
 }
@@ -1289,7 +1289,7 @@ void SetPlayerAvatarStateMask(u8 flags)
     gPlayerAvatar.flags |= flags;
 }
 
-static u8 GetPlayerAvatarStateTransitionByGraphicsId(u8 graphicsId)
+static u8 GetPlayerAvatarStateTransitionByGraphicsId(u16 graphicsId)
 {
     u32 i;
 
@@ -1301,7 +1301,7 @@ static u8 GetPlayerAvatarStateTransitionByGraphicsId(u8 graphicsId)
     return PLAYER_AVATAR_FLAG_ON_FOOT;
 }
 
-u8 GetPlayerAvatarGraphicsIdByCurrentState(void)
+u16 GetPlayerAvatarGraphicsIdByCurrentState(void)
 {
     u32 i;
     u8 flags = gPlayerAvatar.flags;
@@ -1314,7 +1314,7 @@ u8 GetPlayerAvatarGraphicsIdByCurrentState(void)
     return 0;
 }
 
-void SetPlayerAvatarExtraStateTransition(u8 graphicsId, u8 transitionFlag)
+void SetPlayerAvatarExtraStateTransition(u16 graphicsId, u8 transitionFlag)
 {
     u8 stateFlag = GetPlayerAvatarStateTransitionByGraphicsId(graphicsId);
 
@@ -1362,7 +1362,7 @@ void SetPlayerInvisibility(bool8 invisible)
 
 void SetPlayerAvatarAnimation(u32 playerAnimId, u32 animNum)
 {
-    u8 gfxId = sPlayerAvatarAnimGfxIds[playerAnimId][gSaveBlock2Ptr->playerGender];
+    u16 gfxId = sPlayerAvatarAnimGfxIds[playerAnimId][gSaveBlock2Ptr->playerGender];
     ObjectEventSetGraphicsId(&gObjectEvents[gPlayerAvatar.objectEventId], gfxId);
     StartSpriteAnim(&gSprites[gPlayerAvatar.spriteId], animNum);
 }
